# docs/swagger.yaml
definitions:
...
  entity.Instance:
    properties:
      cluster:
        $ref: '#/definitions/model.Cluster'
      cluster_id:
        type: string
      count:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      limit_cpu:
        type: string
      limit_memory:
        type: string
      name:
        type: string
      namespace:
        type: string
      request_cpu:
        type: string
      request_memory:
        type: string
      secret:
        items:
          $ref: '#/definitions/entity.SecretInfo'
        type: array
      service_info:
        $ref: '#/definitions/entity.ServiceInfo'
      service_type:
        type: string
      status:
        type: string
      template:
        $ref: '#/definitions/model.Templates'
      template_id:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
      volume:
        type: string
      yaml:
        type: string
    type: object
  entity.InstancePage:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.Instance'
        type: array
      total:
        type: integer
    type: object
...
  entity.PodCondition:
    properties:
      message:
        type: string
      reason:
        type: string
      status:
        type: string
    type: object
  entity.PodStatus:
    properties:
      conditions:
        items:
          $ref: '#/definitions/entity.PodCondition'
        type: array
      name:
        type: string
      phase:
        type: string
    type: object
...
  entity.SecretInfo:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    type: object
  entity.ServiceAddr:
    properties:
      host:
        type: string
      port:
        type: integer
    type: object
  entity.ServiceInfo:
    properties:
      addresses:
        items:
          $ref: '#/definitions/entity.ServiceAddr'
        type: array
      service_type:
        type: string
    type: object
...
  model.Cluster:
    properties:
      Status:
        type: string
      api_server:
        type: string
      createdAt:
        type: string
      id:
        type: string
      kube_config:
        type: string
      name:
        type: string
      token:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      version:
        type: string
    type: object
  model.Templates:
    properties:
      advance_template:
        type: string
      base_template:
        type: string
      createdAt:
        type: string
      icon:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/instance/:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 实例name
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Instance'
      tags:
      - 根据实例名获取实例
  /api/v1/instances:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 长度
        in: query
        name: size
        required: true
        type: string
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.InstancePage'
      tags:
      - 根据用户获取实例
  /api/v1/instances/:instance_id:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 实例ID
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Instance'
      tags:
      - 根据实例ID获取实例
...
  /api/v1/pod/status:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 集群ID
        in: query
        name: cluster_id
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: pod名称
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.PodStatus'
      tags:
      - 获取Pod的状态
  /api/v1/pods:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 集群ID
        in: query
        name: cluster_id
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      tags:
      - 根据namespace下所有Pod名称
  /api/v1/pods/status:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 集群ID
        in: query
        name: cluster_id
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.PodStatus'
            type: array
      tags:
      - 根据namespace获取其所有Pod的状态
...
swagger: "2.0"